Add a chat input and message list
This commit adds a chat input and message list to the home page.

The chat input is a form with an input and a button. The input is a text input
that allows the user to type a message. The button is a submit button that
sends the message to the server.

The message list is a list of messages that are displayed to the user.

The chat input and message list are both displayed on the home page.