Add Pusher
This commit adds Pusher to the project.

Pusher is a hosted service that makes it super-easy to add realtime data and functionality to web and mobile applications.

This commit adds the Pusher library to the project and uses it to send a message to the server when a new message is sent.