Add Pusher to the app
This commit adds Pusher to the app. It adds a new dependency to the project,
and adds the Pusher client to the client side. It also adds the Pusher server
to the server side.

The server side is used to trigger events when a new message is added. The
client side is used to listen to those events and update the message list
optimistically.